services/correlation/requirements.txt
numpy>=1.26,<2.0
pandas>=2.1,<3.0
psycopg2-binary>=2.9,<3.0
PyYAML>=6.0,<7.0
pytest>=7.4,<9.0
pywebpush>=2.0,<3.0
requests>=2.31,<3.0
responses>=0.25,<1.0
